work on ingredients

+++ recipe-manager-app/src/main/java/com/github/JanBuetikofer/recipemanager/ingredient/IngredientRepo.java
@@ -2,6 +2,6 @@ package com.github.JanBuetikofer.recipemanager.ingredient;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IngredientRepo extends JpaRepository<Ingredient, Integer> {
+public interface IngredientRepo extends JpaRepository<Ingredient, Long> {
     
 }
